+++ temp_era/BetterBlizzFrames.lua
@@ -1291,7 +1291,7 @@ local function executeCustomCode()
     end
 end
 
-function BBF.AddBackgroundTextureToUnitFrames(frame)
+function BBF.AddBackgroundTextureToUnitFrames(frame, tot)
     if not BetterBlizzFramesDB.addUnitFrameBgTexture then
         if frame.bbfBgTexture then
             frame.bbfBgTexture:Hide()
@@ -1308,11 +1308,16 @@ function BBF.AddBackgroundTextureToUnitFrames(frame)
     local bgTex = frame:CreateTexture(nil, "BACKGROUND", nil, -1)
     bgTex:SetColorTexture(unpack(color))
 
-    local topAnchor = frame.healthbar or frame
+    local topAnchor = frame.healthbar or frame.HealthBar or frame
     local bottomAnchor = frame.manabar or frame
 
-    bgTex:SetPoint("TOPLEFT", topAnchor, "TOPLEFT", 0, 0)
-    bgTex:SetPoint("BOTTOMRIGHT", bottomAnchor, "BOTTOMRIGHT", 0, 0)
+    if tot then
+        bgTex:SetPoint("TOPLEFT", topAnchor, "TOPLEFT", -3, 0)
+        bgTex:SetPoint("BOTTOMRIGHT", bottomAnchor, "BOTTOMRIGHT", 0, 0)
+    else
+        bgTex:SetPoint("TOPLEFT", topAnchor, "TOPLEFT", 0, 0)
+        bgTex:SetPoint("BOTTOMRIGHT", bottomAnchor, "BOTTOMRIGHT", 0, 0)
+    end
     if frame.Background then
         frame.Background:Hide()
     end
@@ -1324,6 +1329,10 @@ function BBF.UnitFrameBackgroundTexture()
     BBF.AddBackgroundTextureToUnitFrames(PlayerFrame)
     BBF.AddBackgroundTextureToUnitFrames(TargetFrame)
     --BBF.AddBackgroundTextureToUnitFrames(FocusFrame)
+
+    BBF.AddBackgroundTextureToUnitFrames(TargetFrameToT, true)
+    --BBF.AddBackgroundTextureToUnitFrames(FocusFrameToT, true)
+    BBF.AddBackgroundTextureToUnitFrames(PetFrame, true)
 end
 
 -- Event registration for PLAYER_LOGIN